Refactor transactions into separate files
Separated each transaction function into its own dedicated file for better organization
